requirements.txt
captum==0.2.0
certifi==2023.7.22
cilog==1.3.0
cloudpickle==2.2.1
contourpy==1.1.0
cvxpy==1.3.2
cycler==0.11.0
dill==0.3.7
dive-into-graphs==1.1.0
futures==3.0.5
gdown==4.7.1
geneticpython==0.0.3
googledrivedownloader==0.4
h5py==3.9.0
# Hydra==2.5
hydra-core==1.3.2
isodate==0.6.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.40.1
more-itertools==10.1.0
Mosek==10.1.13
mpmath==1.3.0
munch==4.0.0
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.1
omegaconf==2.3.0
openpyxl==3.1.2
opt-einsum==3.3.0
osqp==0.6.3
packaging==23.1
pandas==2.0.3
patsy==0.5.3
pyparsing==3.1.1
pyscf==2.3.0
qdldl==0.1.7.post0
rdflib==7.0.0
rdkit-pypi==2022.9.5
scikit-learn==1.3.0
scipy==1.10.1
scs==3.2.3
shap==0.42.1
slicer==0.0.7
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.10.0
typed-argument-parser==1.7.2
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
yacs==0.1.8
zipp==3.16.2
gurobipy
scikit-image
lime
ml_collections
optuna
fast_pytorch_kmeans
